+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     the target node, which complicated deployments in non-domain scenarios. Using `PsDscRunAsCredential` solves
     this problems for us.
 
+  - Fixed virtual directory creation for SQL Server 2016.
+
 - Changes to xSQLServerDatabase
   - Changed the readme, SQLInstance should have been SQLInstanceName.
 - Changes to xSQLServerSetup